System::Threading -> Linux Implementation
G++ gives more strict errors than vc.
Consider how to handle int to void*, either with passing a reference to
the original int stack variable, or a pointer like done in this version,
but then the original value is also changed.

If keeping the reference option, create macros VoidPtrTo(Type) and
VoidPtrFrom(Type) to cast.